requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
docker==6.0.1
idna==3.4
packaging==23.0
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
urllib3==1.26.14
websocket-client==1.5.0
